+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All Notable changes to `Slick/i18n` will be documented in this file.
 ## 1.2.0 - 2015-??-??
 
 ### Added
-- No significant or notorious changes were made.
+- Set the translator types as Translator class constants:
+  TYPE_PHP_ARRAY and TYPE_GETTEXT
+- Set class property names convention to PSR-2
 
 ## 1.1.0 - 2015-04-14